Minor Improvements

+++ Mortgage.Api/Application/Dtos/MortgageDto.cs
@@ -1,7 +1,7 @@
 public class MortgageDto
 {
     public string First_Instalment_Date {get; set;} = "";
-    public double Loan_Ammount {get; set;}
-    public double Instalments {get; set;}
-    public double Interest_Rate_In_Percent {get; set;}
+    public decimal Loan_Ammount {get; set;}
+    public int Instalments {get; set;}
+    public decimal Interest_Rate_In_Percent {get; set;}
 }
