+++ Mortgage.Api/Application/Dtos/SnowballDto.cs
@@ -1,5 +1,5 @@
 public class SnowballDto {
-    public double Fixed_Payment {get; set;}
+    public decimal Fixed_Payment {get; set;}
     public string Overpayment_Date {get; set;} = "";
 }
